Renamed objects-memory to objects-memory-amd64 and renamed its objects to OldNameAmd64
