+++ kernel/objects/memory-amd64/mcconf.module
@@ -1,0 +1,17 @@
+# -*- mode:toml; -*-
+
+[module.objects-memory-amd64]
+incfiles = [
+    "objects/FrameDataAmd64.hh",
+    "objects/FrameAmd64.hh", 
+    "objects/MemoryRegion.hh",
+    "objects/PageMapAmd64.hh",
+    "objects/PML4InvalidationBroadcastAmd64.hh",
+    "objects/MemoryRoot.hh",
+    "objects/StaticMemoryRegion.hh" 
+ ]
+kernelfiles = [
+    "objects/MemoryRegion.cc",
+    "objects/PageMapAmd64.cc",
+    "objects/PML4InvalidationBroadcastAmd64.cc"
+]
